L1: PictureViewModel
